refactor: react api 디렉터리 구조 개선
