+++ frontend/rush/src/api/article/CheckHasILikedApi.js
@@ -1,5 +1,5 @@
 import axios from "axios";
-import {BACKEND_ADDRESS} from "../constants/ADDRESS";
+import {BACKEND_ADDRESS} from "../../constants/ADDRESS";
 
 const checkHasILikedApi = (accessToken, articleId, mapType) => {
   const config = {
