+++ frontend/rush/src/api/article/findMyArticlesApi.js
@@ -1,5 +1,5 @@
 import axios from "axios";
-import {BACKEND_ADDRESS} from "../constants/ADDRESS";
+import {BACKEND_ADDRESS} from "../../constants/ADDRESS";
 
 const FindMyArticlesApi = (accessToken) => {
   const config = {
