+++ frontend/rush/src/api/comment/CreateCommentApi.js
@@ -1,5 +1,5 @@
 import axios from "axios";
-import {BACKEND_ADDRESS} from "../constants/ADDRESS";
+import {BACKEND_ADDRESS} from "../../constants/ADDRESS";
 
 const createCommentApi = (content, articleId, mapType, accessToken, history) => {
   const config = {
