+++ frontend/rush/src/api/comment/DeleteCommentApi.js
@@ -1,5 +1,5 @@
 import axios from "axios";
-import {BACKEND_ADDRESS} from "../constants/ADDRESS";
+import {BACKEND_ADDRESS} from "../../constants/ADDRESS";
 
 const deleteCommentApi = ({ commentId, accessToken, history }) => {
   if (!accessToken) {
