+++ frontend/rush/src/api/group/ChangeGroupImportant.js
@@ -1,7 +1,7 @@
 import axios from "axios";
-import {BACKEND_ADDRESS} from "../constants/ADDRESS";
+import {BACKEND_ADDRESS} from "../../constants/ADDRESS";
 
-const deleteArticleApi = ({articleId, accessToken, markerLat, markerLng, history}) => {
+const changeGroupImportant = ({groupId, accessToken, history}) => {
   if (!accessToken) {
     alert("로그인이 필요한 서비스입니다.")
     history.push('/login');
@@ -11,25 +11,18 @@ const deleteArticleApi = ({articleId, accessToken, markerLat, markerLng, history
     headers: {
       Authorization: "Bearer " + accessToken
     }
-  };
-  axios.delete(BACKEND_ADDRESS + "/articles/" + articleId, config)
-  .then(response => {
-    if (response.status === 204) {
-      alert("글이 삭제되었습니다 :)");
-      history.push({
-        pathname:"/",
-        state: {lat: markerLat, lng: markerLng}
-      });
-    }
-  })
+  }
+  axios.put(BACKEND_ADDRESS + "/groups/" + groupId + "/important", null, config)
+  .then(response => {})
   .catch(error => {
     if (error.response.status === 401 || error.response.status === 403) {
       alert("로그인이 만료되었습니다. 다시 로그인해주세요.");
       history.push("/login");
     } else {
-      alert("글 삭제 실패");
+      alert("중요 그룹 표시 토글 실패");
     }
     return Promise.reject();
   });
 };
-export default deleteArticleApi;
+
+export default changeGroupImportant;
