+++ frontend/rush/src/api/group/FindGroupMembersApi.js
@@ -1,7 +1,7 @@
 import axios from "axios";
-import {BACKEND_ADDRESS} from "../constants/ADDRESS";
+import {BACKEND_ADDRESS} from "../../constants/ADDRESS";
 
-const changeGroupImportant = ({groupId, accessToken, history}) => {
+const findGroupMembersApi = ({ groupId, accessToken, history }) => {
   if (!accessToken) {
     alert("로그인이 필요한 서비스입니다.")
     history.push('/login');
@@ -12,17 +12,20 @@ const changeGroupImportant = ({groupId, accessToken, history}) => {
       Authorization: "Bearer " + accessToken
     }
   }
-  axios.put(BACKEND_ADDRESS + "/groups/" + groupId + "/important", null, config)
-  .then(response => {})
+  return axios.get(BACKEND_ADDRESS + "/groups/" + groupId + "/members", config)
+  .then(response => {
+    if (response.status === 200) {
+      return response.data
+    }
+  })
   .catch(error => {
     if (error.response.status === 401 || error.response.status === 403) {
       alert("로그인이 만료되었습니다. 다시 로그인해주세요.");
       history.push("/login");
-    } else {
-      alert("중요 그룹 표시 토글 실패");
+      return Promise.reject();
     }
-    return Promise.reject();
+    alert("이유가 뭔지 모르겠지만 그룹원 목록을 불러오는데 실패했음...");
   });
 };
 
-export default changeGroupImportant;
+export default findGroupMembersApi;
