+++ frontend/rush/src/api/group/JoinGroupApi.js
@@ -1,8 +1,8 @@
-import {ACCESS_TOKEN} from "../constants/SessionStorage";
+import {ACCESS_TOKEN} from "../../constants/SessionStorage";
 import axios from "axios";
-import {BACKEND_ADDRESS} from "../constants/ADDRESS";
+import {BACKEND_ADDRESS} from "../../constants/ADDRESS";
 
-const createGroupApi = ({groupName, history}) => {
+const joinGroupApi = ({invitationCode, history}) => {
   const accessToken = sessionStorage.getItem(ACCESS_TOKEN);
 
   if (!accessToken) {
@@ -10,15 +10,12 @@ const createGroupApi = ({groupName, history}) => {
     history.push('/login');
     return Promise.reject("토큰이 없음");
   }
-  const body = {
-    name: groupName
-  };
   const config = {
     headers: {
       Authorization: "Bearer " + accessToken
     }
   }
-  axios.post(BACKEND_ADDRESS + "/groups", body, config)
+  axios.post(BACKEND_ADDRESS + "/groups/join?invitation_code=" + invitationCode, null, config)
   .then(response => {
     if (response.status === 201) {
       const backGroupUri = response.headers.location;
@@ -32,11 +29,10 @@ const createGroupApi = ({groupName, history}) => {
     if (error.response.status === 401 || error.response.status === 403) {
       alert("로그인이 만료되었습니다. 다시 로그인해주세요.");
       history.push("/login");
-    } else {
-      alert("그룹 가입 실패");
+      return Promise.reject();
     }
-    return Promise.reject();
+    alert("이유가 뭔지 모르겠지만 그룹 가입에 실패했음...");
   });
 };
 
-export default createGroupApi;
+export default joinGroupApi;
