+++ frontend/rush/src/api/group/WithdrawGroupApi.js
@@ -1,5 +1,5 @@
 import axios from "axios";
-import {BACKEND_ADDRESS} from "../constants/ADDRESS";
+import {BACKEND_ADDRESS} from "../../constants/ADDRESS";
 
 const withdrawGroupApi = ({groupId, accessToken, history}) => {
   if (!accessToken) {
