+++ frontend/rush/src/api/user/WithdrawUserApi.js
@@ -1,7 +1,8 @@
+import React from 'react';
 import axios from "axios";
-import {BACKEND_ADDRESS} from "../constants/ADDRESS";
+import {BACKEND_ADDRESS} from "../../constants/ADDRESS";
 
-const editGroupNameApi = ({ groupId, newGroupName, accessToken, history }) => {
+const withdrawUserApi = ({accessToken, history}) => {
   if (!accessToken) {
     alert("로그인이 필요한 서비스입니다.")
     history.push('/login');
@@ -12,19 +13,22 @@ const editGroupNameApi = ({ groupId, newGroupName, accessToken, history }) => {
       Authorization: "Bearer " + accessToken
     }
   };
-  const body = {
-    newName: newGroupName
-  };
-  return axios.put(BACKEND_ADDRESS + "/groups/" + groupId, body, config)
+    axios.delete(BACKEND_ADDRESS + "/users/me", config)
+    .then(response => {
+      if (response.status === 204) {
+        alert("회원 탈퇴가 완료되었습니다");
+        history.push("/");
+      }
+    })
     .catch(error => {
       if (error.response.status === 401 || error.response.status === 403) {
         alert("로그인이 만료되었습니다. 다시 로그인해주세요.");
         history.push("/login");
       } else {
-        alert("그룹이름 변경 실패");
+        alert("회원 탈퇴 실패ㅜ");
       }
       return Promise.reject();
     });
 };
 
-export default editGroupNameApi;
+export default withdrawUserApi;
