+++ frontend/rush/src/components/articleDetail/ArticleBody.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ArticleTitle from "./ArticleTitle";
 import styled from "styled-components";
-import changeMyLikeApi from "../../api/ChangeMyLikeApi";
+import changeMyLikeApi from "../../api/article/ChangeMyLikeApi";
 import DeleteButton from "./DeleteButton";
 
 const ArticleContent = styled.div`
