+++ frontend/rush/src/components/group/GroupPage.js
@@ -2,12 +2,12 @@ import React, {useEffect, useState} from 'react';
 import WindowSize from "../../util/WindowSize";
 import {DisplayBox, Outside} from "../common/Box";
 import {ACCESS_TOKEN} from "../../constants/SessionStorage";
-import findGroupApi from "../../api/FindGroupApi";
+import findGroupApi from "../../api/group/FindGroupApi";
 import {withRouter} from "react-router-dom";
 import GroupHead from "./GroupHead";
 import CancelButton from "./CancelButton";
 import InvitationCode from "./InvitationCode";
-import findGroupMembersApi from "../../api/FindGroupMembersApi";
+import findGroupMembersApi from "../../api/group/FindGroupMembersApi";
 import GroupMembers from "./members/GroupMembers";
 import EditNameModal from "./editNameModal/EditNameModal";
 import WithdrawButton from "./WithdrawButton";
