+++ frontend/rush/src/components/myPage/MyPage.js
@@ -2,14 +2,14 @@ import React, {useEffect, useState} from 'react';
 import {DisplayBox, Outside} from "../common/Box";
 import WindowSize from "../../util/WindowSize";
 import CancelButton from "./CancelButton";
-import findUserApi from "../../api/findUserApi";
-import findMyArticlesApi from "../../api/findMyArticlesApi";
+import findUserApi from "../../api/user/findUserApi";
+import findMyArticlesApi from "../../api/article/findMyArticlesApi";
 import {ACCESS_TOKEN} from "../../constants/SessionStorage";
 import Profile from "./Profile";
 import Info from "./Info";
 import MyArticles from "./articles/MyArticles";
 import MyGroups from "./groups/MyGroups";
-import findMyGroupsApi from "../../api/FindMyGroupsApi";
+import findMyGroupsApi from "../../api/group/FindMyGroupsApi";
 import styled from "styled-components";
 import {withRouter} from "react-router-dom";
 import WithdrawButton from "./WithdrawButton";
